Crate parquet_show_bloom_filter

Binary file to read bloom filter data from a Parquet file.

Install

parquet-show-bloom-filter can be installed using cargo:

cargo install parquet --features=cli

After this parquet-show-bloom-filter should be available:

parquet-show-bloom-filter XYZ.parquet id a

The binary can also be built from the source code and run as follows:

cargo run --features=cli --bin parquet-show-bloom-filter -- --file-name XYZ.parquet --column id --values a

Structs

Args 🔒

Functions

main 🔒